=CHG: DecodeMessageService now takes a byte array as input =CHG: Common EncodeMessageService functionalities from JSON and Protobuf implementation combined
