+++ agrirouter-api-java-api/src/main/kotlin/com/dke/data/agrirouter/api/dto/messaging/inner/MessageResponse.kt
@@ -1,0 +1,5 @@
+package com.dke.data.agrirouter.api.dto.messaging.inner
+
+data class MessageResponse(
+    val message:ByteArray
+)
