+++ agrirouter-api-java-impl/src/main/java/com/dke/data/agrirouter/impl/messaging/encoding/json/DecodeMessageServiceJSONImpl.java
@@ -5,6 +5,7 @@ import com.dke.data.agrirouter.api.dto.encoding.DecodeMessageResponse;
 import com.dke.data.agrirouter.api.exception.CouldNotDecodeMessageException;
 import com.dke.data.agrirouter.api.service.messaging.encoding.DecodeMessageService;
 import com.dke.data.agrirouter.impl.NonEnvironmentalService;
+import com.dke.data.agrirouter.impl.messaging.encoding.DecodeMessageServiceImpl;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
 import java.io.ByteArrayInputStream;
@@ -13,52 +14,19 @@ import java.util.Base64;
 import org.apache.commons.lang3.StringUtils;
 
 /** Internal service implementation. */
-public class DecodeMessageServiceJSONImpl extends NonEnvironmentalService
-    implements DecodeMessageService<String> {
+public class DecodeMessageServiceJSONImpl extends DecodeMessageServiceImpl {
 
-  @Override
+
   public DecodeMessageResponse decode(String encodedResponseString) {
     this.logMethodBegin(encodedResponseString);
 
     if (StringUtils.isBlank(encodedResponseString)) {
       throw new IllegalArgumentException("Please provide a valid encoded response.");
     }
-    try {
 
-      this.getNativeLogger().trace("Decoding byte array.");
-      byte[] decodedBytes = Base64.getDecoder().decode(encodedResponseString);
-      ByteArrayInputStream inputStream = new ByteArrayInputStream(decodedBytes);
 
-      this.getNativeLogger().trace("Parse response envelope.");
-      agrirouter.response.Response.ResponseEnvelope responseEnvelope =
-          agrirouter.response.Response.ResponseEnvelope.parseDelimitedFrom(inputStream);
-
-      this.getNativeLogger().trace("Parse response payload wrapper.");
-      agrirouter.response.Response.ResponsePayloadWrapper responsePayloadWrapper =
-          agrirouter.response.Response.ResponsePayloadWrapper.parseDelimitedFrom(inputStream);
-      DecodeMessageResponse decodeMessageResponse = new DecodeMessageResponse();
-      decodeMessageResponse.setResponseEnvelope(responseEnvelope);
-      decodeMessageResponse.setResponsePayloadWrapper(responsePayloadWrapper);
-
-      this.logMethodEnd(decodeMessageResponse);
-      return decodeMessageResponse;
-    } catch (IOException e) {
-      throw new CouldNotDecodeMessageException(e);
-    }
-  }
-
-  @Override
-  public MessageOuterClass.Message decode(ByteString message) {
-    try {
-      this.logMethodBegin(message);
-
-      this.getNativeLogger().trace("Decoding byte string.");
-      MessageOuterClass.Message decodedMessage = MessageOuterClass.Message.parseFrom(message);
-
-      this.logMethodEnd(decodedMessage);
-      return decodedMessage;
-    } catch (InvalidProtocolBufferException e) {
-      throw new CouldNotDecodeMessageException(e);
-    }
+    this.getNativeLogger().trace("Decoding byte array.");
+    byte[] decodedBytes = Base64.getDecoder().decode(encodedResponseString);
+    return this.decode(decodedBytes);
   }
 }
