+++ agrirouter-api-java-impl/src/main/java/com/dke/data/agrirouter/impl/messaging/encoding/protobuf/EncodeMessageServiceProtobufImpl.java
@@ -8,6 +8,7 @@ import com.dke.data.agrirouter.api.service.parameters.MessageHeaderParameters;
 import com.dke.data.agrirouter.api.service.parameters.PayloadParameters;
 import com.dke.data.agrirouter.api.util.TimestampUtil;
 import com.dke.data.agrirouter.impl.NonEnvironmentalService;
+import com.dke.data.agrirouter.impl.messaging.encoding.EncodeMessageServiceImpl;
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
 import com.sap.iotservices.common.protobuf.gateway.MeasureRequestMessageProtos;
@@ -16,9 +17,9 @@ import java.io.IOException;
 import org.apache.commons.lang3.StringUtils;
 
 /** Internal service implementation. */
-public class EncodeMessageServiceProtobufImpl extends NonEnvironmentalService
-    implements EncodeMessageService {
+public class EncodeMessageServiceProtobufImpl  extends EncodeMessageServiceImpl {
 
+  @Override
   public EncodeMessageResponse.EncodeMessageResponseProtobuf encode(
       MessageHeaderParameters messageHeaderParameters, PayloadParameters payloadParameters) {
     this.logMethodBegin(messageHeaderParameters, payloadParameters);
@@ -29,73 +30,16 @@ public class EncodeMessageServiceProtobufImpl extends NonEnvironmentalService
     messageHeaderParameters.validate();
     payloadParameters.validate();
 
-    try (ByteArrayOutputStream streamedMessage = new ByteArrayOutputStream()) {
+    byte[] messageBuffer = encodeStreamedMessage(messageHeaderParameters, payloadParameters);
 
-      this.getNativeLogger().trace("Encode header.");
-      this.header(messageHeaderParameters).writeDelimitedTo(streamedMessage);
+    MeasureRequestMessageProtos.MeasureRequestMessage.Builder measureRequestBuilder =
+        MeasureRequestMessageProtos.MeasureRequestMessage.newBuilder();
 
-      this.getNativeLogger().trace("Encode payload.");
-      this.payload(payloadParameters).writeDelimitedTo(streamedMessage);
+    measureRequestBuilder.setMessage(ByteString.copyFrom(messageBuffer));
+    MeasureRequestMessageProtos.MeasureRequestMessage measureMessageProtobuf;
+    measureMessageProtobuf = measureRequestBuilder.build();
 
-      this.getNativeLogger().trace("Encoding message.");
-      byte[] encodedByteArray = streamedMessage.toByteArray();
-
-      MeasureRequestMessageProtos.MeasureRequestMessage.Builder measureRequestBuilder =
-          MeasureRequestMessageProtos.MeasureRequestMessage.newBuilder();
-
-      measureRequestBuilder.setMessage(ByteString.copyFrom(encodedByteArray));
-      MeasureRequestMessageProtos.MeasureRequestMessage measureMessageProtobuf;
-      measureMessageProtobuf = measureRequestBuilder.build();
-
-      return new EncodeMessageResponse.EncodeMessageResponseProtobuf(
-          messageHeaderParameters.applicationMessageId, measureMessageProtobuf);
-    } catch (IOException e) {
-      throw new CouldNotEncodeMessageException(e);
-    }
-  }
-
-  private Request.RequestEnvelope header(MessageHeaderParameters parameters) {
-    this.logMethodBegin(parameters);
-
-    this.getNativeLogger().trace("Create message header.");
-    Request.RequestEnvelope.Builder messageHeader = Request.RequestEnvelope.newBuilder();
-    messageHeader.setApplicationMessageId(parameters.getApplicationMessageId());
-    messageHeader.setApplicationMessageSeqNo(parameters.getApplicationMessageSeqNo());
-    messageHeader.setTechnicalMessageType(parameters.getTechnicalMessageType().getKey());
-    messageHeader.setMode(parameters.getMode());
-    if (StringUtils.isNotBlank(parameters.getTeamSetContextId())) {
-      messageHeader.setTeamSetContextId(parameters.getTeamSetContextId());
-    }
-    if (!parameters.getRecipients().isEmpty()) {
-      messageHeader.addAllRecipients(parameters.getRecipients());
-    }
-    if (parameters.getChunkInfo() != null) {
-      messageHeader.setChunkInfo(parameters.getChunkInfo());
-    }
-    messageHeader.setTimestamp(new TimestampUtil().current());
-
-    this.getNativeLogger().trace("Build message envelope.");
-    Request.RequestEnvelope requestEnvelope = messageHeader.build();
-
-    this.logMethodEnd(requestEnvelope);
-    return requestEnvelope;
-  }
-
-  private Request.RequestPayloadWrapper payload(PayloadParameters parameters) {
-    this.logMethodBegin(parameters);
-
-    this.getNativeLogger().trace("Create message payload.");
-    Request.RequestPayloadWrapper.Builder messagePayload =
-        Request.RequestPayloadWrapper.newBuilder();
-    Any.Builder builder = Any.newBuilder();
-    builder.setTypeUrl(parameters.getTypeUrl());
-    builder.setValue(parameters.getValue());
-    messagePayload.setDetails(builder.build());
-
-    this.getNativeLogger().trace("Message message payload wrapper.");
-    Request.RequestPayloadWrapper requestPayloadWrapper = messagePayload.build();
-
-    this.logMethodEnd(requestPayloadWrapper);
-    return requestPayloadWrapper;
+    return new EncodeMessageResponse.EncodeMessageResponseProtobuf(
+        messageHeaderParameters.applicationMessageId, measureMessageProtobuf);
   }
 }
